refactor(api): rename/move app-config to feature-flags
Ref. metriport/metriport-internal#2840
